refactor :  ResponseDto - Object 사용 및 위치 controller 패키지의 dto로 이동. ErrorResponseDto - Dto로 이름 변경 및 data컬럼 ErrorData클래스로 반환타입 변경.
