+++ src/main/java/com/starter/performance/controller/dto/ResponseDto.java
@@ -1,4 +1,4 @@
-package com.starter.performance.service.dto;
+package com.starter.performance.controller.dto;
 
 import lombok.AccessLevel;
 import lombok.Builder;
