+++ src/main/java/com/starter/performance/exception/AbstractException.java
@@ -4,6 +4,6 @@ public abstract class AbstractException extends RuntimeException {
 
     abstract public int getStatusCode();
 
-    abstract public Object getData();
+    abstract public ErrorData getData();
 
 }
