+++ src/main/java/com/starter/performance/exception/ErrorData.java
@@ -1,0 +1,12 @@
+package com.starter.performance.exception;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@Getter
+@RequiredArgsConstructor
+public class ErrorData {
+
+    private final String errorType;
+
+}
