+++ src/main/java/com/starter/performance/exception/dto/ErrorResponseDto.java
@@ -1,5 +1,6 @@
-package com.starter.performance.exception;
+package com.starter.performance.exception.dto;
 
+import com.starter.performance.exception.ErrorData;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -7,13 +8,13 @@ import lombok.NoArgsConstructor;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ErrorResponse {
+public class ErrorResponseDto {
 
     private int statusCode;
-    private Object data;
+    private ErrorData data;
 
     @Builder
-    public ErrorResponse(int statusCode, Object data) {
+    public ErrorResponseDto(int statusCode, ErrorData data) {
         this.statusCode = statusCode;
         this.data = data;
     }
