+++ src/main/java/com/starter/performance/exception/impl/NotMatchReservationAndMemberException.java
@@ -1,6 +1,7 @@
 package com.starter.performance.exception.impl;
 
 import com.starter.performance.exception.AbstractException;
+import com.starter.performance.exception.ErrorData;
 import com.starter.performance.exception.ReviewErrorCode;
 import org.springframework.http.HttpStatus;
 
@@ -12,7 +13,7 @@ public class NotMatchReservationAndMemberException extends AbstractException {
     }
 
     @Override
-    public Object getData() {
-        return ReviewErrorCode.NOT_MATCH_RESERVATION_AND_MEMBER_EXCEPTION;
+    public ErrorData getData() {
+        return new ErrorData(ReviewErrorCode.NOT_MATCH_RESERVATION_AND_MEMBER_EXCEPTION.name());
     }
 }
