+++ src/main/java/com/starter/performance/service/ReviewService.java
@@ -1,9 +1,9 @@
 package com.starter.performance.service;
 
+import com.starter.performance.controller.dto.ResponseDto;
 import com.starter.performance.controller.dto.ReviewRequestDto;
 import com.starter.performance.domain.Member;
 import com.starter.performance.domain.Reservation;
-import com.starter.performance.service.dto.ResponseDto;
 import org.springframework.security.core.Authentication;
 
 public interface ReviewService {
