+++ src/main/java/com/starter/performance/service/impl/ReviewServiceImpl.java
@@ -1,5 +1,6 @@
 package com.starter.performance.service.impl;
 
+import com.starter.performance.controller.dto.ResponseDto;
 import com.starter.performance.controller.dto.ReviewRequestDto;
 import com.starter.performance.domain.Member;
 import com.starter.performance.domain.PerformanceStatus;
@@ -12,7 +13,6 @@ import com.starter.performance.repository.MemberRepository;
 import com.starter.performance.repository.ReservationRepository;
 import com.starter.performance.repository.ReviewRepository;
 import com.starter.performance.service.ReviewService;
-import com.starter.performance.service.dto.ResponseDto;
 import com.starter.performance.service.dto.ReviewResponseDto;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
